Check that dependencies are declared consistently at source set level
Uklib spec prohibits uklibs from filtering their dependencies and
assumes that uklibs are going to be published in a world of library to
library dependencies. To simulate this in KGP we check that dependencies
are declared consistently across compilations that are going to be
published as Uklib fragments.

This check is inaccurate. For example platform and metadata compilations
might resolve to different versions of the same dependency.

^KT-70493