[Gradle] Expose adhocSoftwareComponent in KMP Extension
With `adhocSoftwareComponent` users can add additional variants
to the "kotlinMultiplatform" root publication. For their own needs.

This implementation follows Gradle's recommendation to use
AdhocSoftwareComponent API to customize publications.

However, there is a difference with Java Gradle Plugin.
As they recommend to cast
components["java"] to AdhocComponentWithVariants
and then add extra variants or change existing.

In KMP we can't do that. because we don't control
AdhocComponentWithVariants interface and cant guarantee stability by
extending the interface because it can change in time.
So exposing a delegate is a good tradeoff.

^KT-58830 Verification Pending