[FIR2IR] Set origin for implicit receivers to IMPLICIT_ARGUMENT
The compiler frontend tracks which receiver - if any - is specified
explicitly. This information is not maintained when converted to IR.
This means a plugin like power-assert is unable to ignore arguments
which are supplied implicitly.

Add a new IrStatementOrigin which indicates when an argument is
implicit. This can be used by power-assert to determine what arguments
can be ignored and to also fix alignment of some operator calls.

^KT-73897